Organize headers into module directories
