+++ src/control/state_handler_node.cpp
@@ -1,6 +1,6 @@
 #include <functional>
 #include <chrono>
-#include "robofer/state_handler.hpp"
+#include "robofer/control/state_handler.hpp"
 
 using namespace std::chrono_literals;
