+++ src/screen/display.cpp
@@ -1,4 +1,4 @@
-#include "robofer/display.hpp"
+#include "robofer/screen/display.hpp"
 
 #include <opencv2/highgui.hpp>
 #include <opencv2/imgproc.hpp>
