+++ src/screen/eyes.cpp
@@ -1,4 +1,4 @@
-#include "robofer/eyes.hpp"
+#include "robofer/screen/eyes.hpp"
 #include <opencv2/imgproc.hpp>
 #include <algorithm>
