+++ src/screen/eyes_st7735_node.cpp
@@ -9,7 +9,7 @@
 #include <cstring>
 #include <vector>
 #include <algorithm>
-#include "robofer/eyes.hpp"
+#include "robofer/screen/eyes.hpp"
 
 using robo_eyes::RoboEyes;
 using robo_eyes::Mood;
